# Backend/api_documentation.yaml
openapi: 3.1.0
info:
  title: CivicResolve AI
  version: 0.1.0
paths:
...
  /api/notifications/{notif_id}/read:
    patch:
      tags:
      - notifications
      summary: Mark As Read
      operationId: mark_as_read_api_notifications__notif_id__read_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: notif_id
        in: path
        required: true
        schema:
          type: integer
          title: Notif Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/notifications/{notif_id}/unread:
    patch:
      tags:
      - notifications
      summary: Mark As Unread
      operationId: mark_as_unread_api_notifications__notif_id__unread_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: notif_id
        in: path
        required: true
        schema:
          type: integer
          title: Notif Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/officer/dash:
    get:
      tags:
      - officer
      summary: Officer Dashboard
      operationId: officer_dashboard_api_officer_dash_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficerDashboardResponse'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    ComplaintSchema:
      properties:
        id:
          type: integer
          title: Id
        token:
          type: string
          title: Token
        relatedTokens:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Relatedtokens
        assignedOfficerId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assignedofficerid
        assignedOfficerName:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignedofficername
        departmentId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Departmentid
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        submittedPhotos:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Submittedphotos
        status:
          type: string
          title: Status
        severity:
          type: string
          title: Severity
        resolutionPhotos:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Resolutionphotos
        resolutionNote:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolutionnote
        createdAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Createdat
        updatedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updatedat
        resolvedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Resolvedat
        closedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Closedat
      type: object
      required:
      - id
      - token
      - title
      - description
      - status
      - severity
      title: ComplaintSchema
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NotificationSchema:
      properties:
        id:
          type: integer
          title: Id
        userId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Userid
        title:
          type: string
          title: Title
        message:
          type: string
          title: Message
        notifType:
          type: string
          title: Notiftype
        isRead:
          type: boolean
          title: Isread
        createdAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Createdat
      type: object
      required:
      - id
      - title
      - message
      - notifType
      - isRead
      title: NotificationSchema
    OfficerDashboardResponse:
      properties:
        officerName:
          type: string
          title: Officername
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        jurisdictionZone:
          anyOf:
          - type: string
          - type: 'null'
          title: Jurisdictionzone
        assignedTickets:
          items:
            $ref: '#/components/schemas/ComplaintSchema'
          type: array
          title: Assignedtickets
        totalAssigned:
          type: integer
          title: Totalassigned
      type: object
      required:
      - officerName
      - assignedTickets
      - totalAssigned
      title: OfficerDashboardResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer